Moved eeprom stuff to it's own file saveData
