+++ OASMan_ESP32/src/main.cpp
@@ -1,8 +1,5 @@
 // OASMan ESP32
 
-#include <Arduino.h>
-#include <EEPROM.h>
-
 #include "user_defines.h"
 #include "input_type.h"
 #include "solenoid.h"
@@ -11,6 +8,7 @@
 #include "compressor.h"
 #include "bitmaps.h"
 #include "bt.h"
+#include "saveData.h"
 
 #if SCREEN_MOODE == true
 #include <SPI.h>
@@ -28,82 +26,7 @@ Manifold *getManifold() {
 }
 //InputType *manifoldSafetyWire;
 
-struct Profile {
-  byte pressure[4];
-};
 
-struct EEPROM_DATA {
-  byte riseOnStart;
-  byte baseProfile;
-  byte raiseOnPressure;
-  Profile profile[MAX_PROFILE_COUNT];
-} EEPROM_DATA;
-#define EEPROM_SIZE sizeof(EEPROM_DATA)
-
-void saveEEPROM() {
-  EEPROM.put(0, EEPROM_DATA);
-  EEPROM.commit();
-}
-void beginEEPROM() {
-  EEPROM.begin(EEPROM_SIZE);
-  EEPROM.get(0, EEPROM_DATA);
-}
-
-byte currentProfile[4];
-bool sendProfileBT = false;
-
-void readProfile(byte profileIndex) {
-  currentProfile[WHEEL_FRONT_PASSENGER] = EEPROM_DATA.profile[profileIndex].pressure[WHEEL_FRONT_PASSENGER];
-  currentProfile[WHEEL_REAR_PASSENGER] = EEPROM_DATA.profile[profileIndex].pressure[WHEEL_REAR_PASSENGER];
-  currentProfile[WHEEL_FRONT_DRIVER] = EEPROM_DATA.profile[profileIndex].pressure[WHEEL_FRONT_DRIVER];
-  currentProfile[WHEEL_REAR_DRIVER] = EEPROM_DATA.profile[profileIndex].pressure[WHEEL_REAR_DRIVER];
-  sendProfileBT = true;
-}
-
-void writeProfile(byte profileIndex) {
-
-  if (currentProfile[WHEEL_FRONT_PASSENGER] != EEPROM_DATA.profile[profileIndex].pressure[WHEEL_FRONT_PASSENGER] ||
-      currentProfile[WHEEL_REAR_PASSENGER] != EEPROM_DATA.profile[profileIndex].pressure[WHEEL_REAR_PASSENGER] ||
-      currentProfile[WHEEL_FRONT_DRIVER] != EEPROM_DATA.profile[profileIndex].pressure[WHEEL_FRONT_DRIVER] ||
-      currentProfile[WHEEL_REAR_DRIVER] != EEPROM_DATA.profile[profileIndex].pressure[WHEEL_REAR_DRIVER]) {
-
-    EEPROM_DATA.profile[profileIndex].pressure[WHEEL_FRONT_PASSENGER] = currentProfile[WHEEL_FRONT_PASSENGER];
-    EEPROM_DATA.profile[profileIndex].pressure[WHEEL_REAR_PASSENGER] = currentProfile[WHEEL_REAR_PASSENGER];
-    EEPROM_DATA.profile[profileIndex].pressure[WHEEL_FRONT_DRIVER] = currentProfile[WHEEL_FRONT_DRIVER];
-    EEPROM_DATA.profile[profileIndex].pressure[WHEEL_REAR_DRIVER] = currentProfile[WHEEL_REAR_DRIVER];
-    saveEEPROM();
-  }
-}
-
-bool getRiseOnStart() {
-    return EEPROM_DATA.riseOnStart;
-}
-void setRiseOnStart(bool value) {
-  if (getRiseOnStart() != value) {
-    EEPROM_DATA.riseOnStart = value;
-    saveEEPROM();
-  }
-}
-
-byte getBaseProfile() {
-    return EEPROM_DATA.baseProfile;
-}
-void setBaseProfile(byte value) {
-  if (getBaseProfile() != value) {
-    EEPROM_DATA.baseProfile = value;
-    saveEEPROM();
-  }
-}
-
-bool getRaiseOnPressureSet() {
-    return EEPROM_DATA.raiseOnPressure;
-}
-void setRaiseOnPressureSet(bool value) {
-  if (getRaiseOnPressureSet() != value) {
-    EEPROM_DATA.raiseOnPressure = value;
-    saveEEPROM();
-  }
-}
 
 Compressor *compressor;
 Wheel *wheel[4];
@@ -290,6 +213,8 @@ void setup() {
       airUp();
     }
   #endif
+
+  Serial.println(F("Startup Complete"));
 }
